refactor: move errors into the utils folder
